+++ src/cli.rs
@@ -1,0 +1,25 @@
+use crate::error::{Error, StandardIOAction};
+use std::io::Write;
+
+pub fn ask(question: &str) -> Result<String, Error> {
+    print!("{}: ", question);
+    std::io::stdout().flush().unwrap();
+    let mut answer = String::new();
+    std::io::stdin()
+        .read_line(&mut answer)
+        .map_err(|e| Error::StandardIO {
+            action: StandardIOAction::Read,
+            err: Some(e.kind()),
+        })?;
+    Ok(answer.trim().to_string())
+}
+
+pub fn ask_password(question: &str) -> Result<String, Error> {
+    let prompt = format!("{} (will not be printed as you type): ", question);
+    rpassword::read_password_from_tty(Some(prompt.as_str()))
+        .map_err(|e| Error::StandardIO {
+            action: StandardIOAction::Read,
+            err: Some(e.kind()),
+        })
+        .map(|s| s.trim().to_string())
+}