+++ src/format/command.rs
@@ -1,26 +1,38 @@
-use crate::error::{Error, FileIOAction, FileKind, StandardIOAction};
-use std::fs::File;
-use std::io::{Read, Write};
-use std::path::PathBuf;
-use std::str::FromStr;
+use crate::{
+    error::{Error, FileIOAction, FileKind, StandardIOAction},
+    file,
+};
+use std::{
+    fs::File,
+    io::{Read, Write},
+    path::PathBuf,
+    str::FromStr,
+};
 
 #[derive(Debug, PartialEq)]
 pub struct Formatted {
-    pub path: PathBuf,
+    pub source: PathBuf,
+    pub destination: Destination,
     pub input: String,
     pub output: String,
 }
 
+#[derive(Debug, PartialEq)]
+pub enum Destination {
+    Stdout,
+    File { path: PathBuf },
+}
+
 pub fn run(stdin: bool, check: bool, files: Vec<String>) -> Result<(), Error> {
-    if stdin {
-        format_stdin(check)
+    let formatted = if stdin {
+        vec![read_and_format_stdin()?]
     } else {
-        let formatted = read_and_format_paths(files)?;
-        if check {
-            check_formatting(formatted)
-        } else {
-            write_formatted(formatted)
-        }
+        read_and_format_paths(files)?
+    };
+    if check {
+        check_formatting(formatted)
+    } else {
+        write_formatted(formatted)
     }
 }
 
@@ -39,21 +51,34 @@ fn check_formatting(formatted_files: Vec<Formatted>) -> Result<(), Error> {
 
 fn write_formatted(formatted_files: Vec<Formatted>) -> Result<(), Error> {
     for formatted in formatted_files {
-        let path = formatted.path;
-        let mut f = File::open(&path).map_err(|e| Error::FileIO {
-            action: FileIOAction::Create,
-            kind: FileKind::File,
-            path: path.clone(),
-            err: Some(e.to_string()),
-        })?;
+        match formatted {
+            Formatted {
+                destination: Destination::Stdout,
+                output,
+                ..
+            } => {
+                print!("{}", output);
+            }
+            Formatted {
+                destination: Destination::File { path },
+                output,
+                ..
+            } => {
+                let mut f = File::create(&path).map_err(|e| Error::FileIO {
+                    action: FileIOAction::Create,
+                    kind: FileKind::File,
+                    path: path.clone(),
+                    err: Some(e.to_string()),
+                })?;
 
-        f.write_all(formatted.output.as_bytes())
-            .map_err(|e| Error::FileIO {
-                action: FileIOAction::WriteTo,
-                kind: FileKind::File,
-                path: path.clone(),
-                err: Some(e.to_string()),
-            })?;
+                f.write_all(output.as_bytes()).map_err(|e| Error::FileIO {
+                    action: FileIOAction::WriteTo,
+                    kind: FileKind::File,
+                    path: path.clone(),
+                    err: Some(e.to_string()),
+                })?;
+            }
+        }
     }
 
     Ok(())
@@ -71,7 +96,7 @@ pub fn read_and_format_paths(files: Vec<String>) -> Result<Vec<Formatted>, Error
         })?;
 
         if path.is_dir() {
-            for path in crate::project::gleam_files(&path).into_iter() {
+            for path in file::gleam_files(&path).into_iter() {
                 formatted_files.push(format_file(path)?);
             }
         } else {
@@ -97,13 +122,14 @@ fn format_file(path: PathBuf) -> Result<Formatted, Error> {
     })?;
 
     Ok(Formatted {
-        path,
+        source: path.clone(),
+        destination: Destination::File { path },
         input: src,
         output: formatted,
     })
 }
 
-pub fn format_stdin(_check: bool) -> Result<(), Error> {
+pub fn read_and_format_stdin() -> Result<Formatted, Error> {
     let mut src = String::new();
     std::io::stdin()
         .read_to_string(&mut src)
@@ -114,10 +140,13 @@ pub fn format_stdin(_check: bool) -> Result<(), Error> {
 
     let formatted = crate::format::pretty(src.as_ref()).map_err(|error| Error::Parse {
         path: PathBuf::from("<standard input>"),
+        src: src.clone(),
         error,
-        src,
     })?;
-
-    print!("{}", formatted);
-    Ok(())
+    Ok(Formatted {
+        source: PathBuf::from("<standard input>"),
+        destination: Destination::Stdout,
+        input: src,
+        output: formatted,
+    })
 }