+++ src/main.rs
@@ -1,10 +1,12 @@
 #![deny(warnings)]
 
 mod ast;
+mod cli;
 mod diagnostic;
-mod doc;
+mod docs;
 mod erl;
 mod error;
+mod file;
 mod format;
 mod new;
 mod parser;
@@ -34,12 +36,7 @@ extern crate lalrpop_util;
 #[macro_use]
 extern crate lazy_static;
 
-use crate::{
-    error::Error,
-    project::{ModuleOrigin, OutputFile, ProjectConfig},
-};
-use std::fs::File;
-use std::io::{Read, Write};
+use crate::error::Error;
 use std::path::PathBuf;
 use structopt::clap::AppSettings;
 use structopt::StructOpt;
@@ -53,10 +50,11 @@ enum Command {
     #[structopt(name = "build", about = "Compile a project")]
     Build {
         #[structopt(help = "location of the project root", default_value = ".")]
-        path: String,
-        #[structopt(help = "generate docs for this package as well", long)]
-        doc: bool,
+        project_root: String,
     },
+
+    #[structopt(name = "docs", about = "Render HTML documentation for a project")]
+    Docs(Docs),
 
     #[structopt(name = "new", about = "Create a new project")]
     New {
@@ -67,7 +65,7 @@ enum Command {
         description: Option<String>,
 
         #[structopt(help = "location of the project root")]
-        path: Option<String>,
+        project_root: Option<String>,
 
         #[structopt(
             long = "template",
@@ -98,22 +96,61 @@ enum Command {
     },
 }
 
+#[derive(StructOpt, Debug)]
+enum Docs {
+    #[structopt(name = "build", about = "Render HTML docs locally")]
+    Build {
+        #[structopt(help = "location of the project root", default_value = ".")]
+        project_root: String,
+
+        #[structopt(help = "the directory to write the docs to", long = "to")]
+        to: Option<String>,
+    },
+
+    #[structopt(name = "publish", about = "Publish HTML docs to HexDocs")]
+    Publish {
+        #[structopt(help = "location of the project root", default_value = ".")]
+        project_root: String,
+
+        #[structopt(help = "the version to publish ah", long = "version")]
+        version: String,
+    },
+
+    #[structopt(name = "remove", about = "Remove HTML docs from HexDocs")]
+    Remove {
+        #[structopt(help = "the name of the package", long = "package")]
+        package: String,
+
+        #[structopt(help = "the version to remove the docs of", long = "version")]
+        version: String,
+    },
+}
+
 fn main() {
     let result = match Command::from_args() {
-        Command::Build { path, doc } => command_build(path, doc),
+        Command::Build { project_root } => command_build(project_root),
+
+        Command::Docs(Docs::Build { project_root, to }) => docs::command::build(project_root, to),
+
+        Command::Docs(Docs::Publish {
+            project_root,
+            version,
+        }) => docs::command::publish(project_root, version),
+
+        Command::Docs(Docs::Remove { package, version }) => docs::command::remove(package, version),
 
         Command::Format {
             stdin,
             files,
             check,
-        } => crate::format::command::run(stdin, check, files),
+        } => format::command::run(stdin, check, files),
 
         Command::New {
             name,
             description,
-            path,
+            project_root,
             template,
-        } => crate::new::create(template, name, description, path, VERSION),
+        } => new::create(template, name, description, project_root, VERSION),
     };
 
     if let Err(e) = result {
@@ -122,141 +159,25 @@ fn main() {
     }
 }
 
-fn command_build(root: String, write_docs: bool) -> Result<(), Error> {
-    let mut srcs = vec![];
+fn command_build(root: String) -> Result<(), Error> {
+    let root = PathBuf::from(&root);
 
-    // Read gleam.toml
-    let project_config = read_project_config(&root)?;
+    // Read and type check project
+    let (_config, analysed) = project::read_and_analyse(&root)?;
 
-    let root_path = PathBuf::from(&root);
-    let lib_dir = root_path.join("_build").join("default").join("lib");
-    let checkouts_dir = root_path.join("_checkouts");
+    // Generate Erlang code
+    let output_files = erl::generate_erlang(analysed.as_slice());
 
-    for project_dir in [lib_dir, checkouts_dir]
-        .iter()
-        .filter_map(|d| std::fs::read_dir(d).ok())
-        .flat_map(|d| d.filter_map(Result::ok))
-        .map(|d| d.path())
-        .filter(|p| {
-            p.file_name().and_then(|os_string| os_string.to_str()) != Some(&project_config.name)
-        })
-    {
-        crate::project::collect_source(
-            project_dir.join("src"),
-            ModuleOrigin::Dependency,
-            &mut srcs,
-        )?;
-    }
-
-    // Collect source code from top level project
-    crate::project::collect_source(root_path.join("src"), ModuleOrigin::Src, &mut srcs)?;
-    crate::project::collect_source(root_path.join("test"), ModuleOrigin::Test, &mut srcs)?;
-
-    let analysed = crate::project::analysed(srcs)?;
-
-    // Generate outputs (Erlang code, html documentation, etc)
-    let mut output_files = vec![];
-    if write_docs {
-        let dir = root_path.join("docs");
-        crate::doc::generate_html(
-            &project_config,
-            analysed.as_slice(),
-            &mut output_files,
-            &dir,
-        );
-        delete_dir(&dir)?;
-    } else {
-        let dir = root_path.join("gen");
-        crate::project::generate_erlang(analysed.as_slice(), &mut output_files);
-        delete_dir(&dir)?;
-    }
-
-    // Delete the gen directory before generating the newly compiled files
-    for file in output_files {
-        write_file(file)?;
-    }
+    // Reset output directory
+    file::delete_dir(&root.join(project::OUTPUT_DIR_NAME))?;
 
     // Print warnings
-    for a in analysed.iter() {
-        for w in a.warnings.iter() {
-            w.pretty_print()
-        }
-    }
+    warning::print_all(analysed.as_slice());
+
+    // Delete the gen directory before generating the newly compiled files
+    file::write_outputs(output_files.as_slice())?;
 
     println!("Done!");
 
     Ok(())
-}
-
-fn delete_dir(dir: &PathBuf) -> Result<(), Error> {
-    if dir.exists() {
-        std::fs::remove_dir_all(&dir).map_err(|e| Error::FileIO {
-            action: error::FileIOAction::Delete,
-            kind: error::FileKind::Directory,
-            path: dir.clone(),
-            err: Some(e.to_string()),
-        })?;
-    }
-    Ok(())
-}
-
-pub fn write_file(file: OutputFile) -> Result<(), Error> {
-    let OutputFile { path, text } = file;
-
-    let dir_path = path.parent().ok_or_else(|| Error::FileIO {
-        action: error::FileIOAction::FindParent,
-        kind: error::FileKind::Directory,
-        path: path.clone(),
-        err: None,
-    })?;
-
-    std::fs::create_dir_all(dir_path).map_err(|e| Error::FileIO {
-        action: error::FileIOAction::Create,
-        kind: error::FileKind::Directory,
-        path: dir_path.to_path_buf(),
-        err: Some(e.to_string()),
-    })?;
-
-    let mut f = File::create(&path).map_err(|e| Error::FileIO {
-        action: error::FileIOAction::Create,
-        kind: error::FileKind::File,
-        path: path.clone(),
-        err: Some(e.to_string()),
-    })?;
-
-    f.write_all(text.as_bytes()).map_err(|e| Error::FileIO {
-        action: error::FileIOAction::WriteTo,
-        kind: error::FileKind::File,
-        path: path.clone(),
-        err: Some(e.to_string()),
-    })?;
-    Ok(())
-}
-
-fn read_project_config(root: &str) -> Result<ProjectConfig, Error> {
-    let config_path = PathBuf::from(root).join("gleam.toml");
-
-    let mut file = File::open(&config_path).map_err(|e| Error::FileIO {
-        action: error::FileIOAction::Open,
-        kind: error::FileKind::File,
-        path: config_path.clone(),
-        err: Some(e.to_string()),
-    })?;
-
-    let mut toml = String::new();
-    file.read_to_string(&mut toml).map_err(|e| Error::FileIO {
-        action: error::FileIOAction::Read,
-        kind: error::FileKind::File,
-        path: config_path.clone(),
-        err: Some(e.to_string()),
-    })?;
-
-    let project_config = toml::from_str(&toml).map_err(|e| Error::FileIO {
-        action: error::FileIOAction::Parse,
-        kind: error::FileKind::File,
-        path: config_path.clone(),
-        err: Some(e.to_string()),
-    })?;
-
-    Ok(project_config)
 }